+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/socket/ChatSocket.kt
@@ -268,6 +268,9 @@ internal open class ChatSocket constructor(
             ChatErrorCode.API_KEY_NOT_FOUND.code,
             ChatErrorCode.VALIDATION_ERROR.code,
             -> {
+                logger.d {
+                    "One unrecoverable error happened. Error: ${error.stringify()}. Error code: ${error.streamCode}"
+                }
                 disconnect(DisconnectCause.UnrecoverableError(error))
             }
             else -> {