+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/QueryChannelsImplRepositoryTest.kt
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-@file:Suppress("DEPRECATION_ERROR")
-
 package io.getstream.chat.android.offline.repository
 
 import io.getstream.chat.android.client.api.models.ContainsFilterObject
 import io.getstream.chat.android.client.api.models.NeutralFilterObject
-import io.getstream.chat.android.client.api.models.QuerySort
+import io.getstream.chat.android.client.api.models.querysort.QuerySortByField
 import io.getstream.chat.android.client.models.Filters
 import io.getstream.chat.android.client.test.randomQueryChannelsSpec
 import io.getstream.chat.android.offline.randomQueryChannelsEntity
@@ -77,11 +75,11 @@ internal class QueryChannelsImplRepositoryTest {
         whenever(dao.select(any())) doReturn randomQueryChannelsEntity(
             id = "id1",
             filter = Filters.contains("cid", "cid1"),
-            querySort = QuerySort(),
+            querySort = QuerySortByField(),
             cids = listOf("cid1")
         )
 
-        val result = sut.selectBy(Filters.contains("cid", "cid1"), QuerySort())
+        val result = sut.selectBy(Filters.contains("cid", "cid1"), QuerySortByField())
 
         result.shouldNotBeNull()
         result.filter.shouldBeInstanceOf<ContainsFilterObject>()
@@ -94,7 +92,7 @@ internal class QueryChannelsImplRepositoryTest {
     fun `Given no row in DB with such id When select by id Should return null`() = runTest {
         whenever(dao.select(any())) doReturn null
 
-        val result = sut.selectBy(NeutralFilterObject, QuerySort())
+        val result = sut.selectBy(NeutralFilterObject, QuerySortByField())
 
         result.shouldBeNull()
     }